requirements.txt
asgiref==3.6.0
certifi==2022.12.7
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
pip-tools==6.12.1
pipreqs==0.4.11
pymongo==3.12.0
sqlparse==0.2.4
